refactor(10i.0-LP.4): coverage parser dispatch via pack ownership
Closes audit item #10 from `tmp/phase-10i-LP-roadmap.md`.

Each language pack now owns its coverage parser. Pre-LP.4, the four
JSON-shape parsers (Istanbul summary, Istanbul final, coverage.py,
Go cover-profile) all lived in `src/analyzers/tools/coverage.ts` and
were imported back into the pack files that consumed them. Cobertura
(.NET / Rust) and lcov (Rust) already lived in their packs; this
commit extends that pattern to the remaining four formats:

  parseIstanbulSummary  (typescript) — moved to src/languages/typescript.ts
  parseIstanbulFinal    (typescript) — moved to src/languages/typescript.ts
  parseCoveragePy       (python)     — moved to src/languages/python.ts
  parseGoCoverProfile   (go)         — moved to src/languages/go.ts
  resolveGoPath         (go helper)  — moved to src/languages/go.ts

`src/analyzers/tools/coverage.ts` is now ~80 lines (was 298). It
retains:
  - `Coverage`, `FileCoverage`, `CoverageSource` shared types
  - `loadCoverage(cwd)` capability dispatcher (unchanged — already
    iterated `detectActiveLanguages(cwd)` and pulled each pack's
    `capabilities.coverage` provider)
  - `round1`, `toRelative` shared utilities, now exported for the
    parsers in pack files to import

Skipped: adding a `LanguageSupport.coverageParser?: (raw, path, cwd)
=> Coverage | null` field as the roadmap originally suggested. Adding
it would be a no-caller field — the existing `capabilities.coverage`
provider (via `gather(cwd)`) already does the dispatching cleanly,
and adding a sibling sync entry-point with no consumer is overhead.
If a future caller needs raw-bytes-to-Coverage parsing (e.g. a
`vyuh-dxkit coverage-info <file>` CLI command), it can pull the
parser directly from the pack module. Lower-cost optionality.

Test imports (`test/coverage.test.ts`) updated to import each parser
from its new pack-file home.

Audit progress: 10/14 items closed (1-7, 10, 11, 12 cumulative).
Remaining for this phase: items #9 (LP.5 — detect.ts iteration), #13
(LP.6 — constants.ts pack metadata). Items #8, #14 stay deferred to
10f.4 / future phases.

Regression: 814/814 in 2:07.
